+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
+## 27.2 - 2020/10/18
+- Added sarif format https://sarifweb.azurewebsites.net/
+- Updated rule ids to be in the form `PYG.[db].[index]` from previous `[db][index]`
+
+
 ## 27.1.1 - 2020/10/14
 - Bugfixes